+++ GH.DD.ConfigRetriever/ILogger.cs
@@ -1,0 +1,7 @@
+﻿namespace GH.DD.ConfigRetriever
+{
+    public interface ILogger
+    {
+        void Debug(string message);
+    }
+}