+++ GH.DD.ConfigRetriever/Loggers/StdoutLogger.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace GH.DD.ConfigRetriever.Loggers
+{
+    public class StdoutLogger : ILogger
+    {
+        public void Debug(string message)
+        {
+            Console.WriteLine(message);
+        }
+    }
+}