+++ GH.DD.ConfigRetriever/Loggers/StubLogger.cs
@@ -1,0 +1,8 @@
+﻿namespace GH.DD.ConfigRetriever.Loggers
+{
+    public class StubLogger : ILogger
+    {
+        public void Debug(string message)
+        { }
+    }
+}